requirement.txt
langgraph 
langchain-google-genai 
langchain-core
opencv-python 
 
yt-dlp
pydantic
python-dotenv

#streamlit
youtube-transcript-api

fastapi
uvicorn

google-api-python-client

ipython